--- src/main/java/sakura/task/SakuraException.java
@@ -1,15 +1,0 @@
-package sakura.task;
-
-/**
- * Custom exception for Sakura bot
- */
-public class SakuraException extends Exception {
-    /**
-     * Constructs a SakuraException Class
-     *
-     * @param message the message as a string
-     */
-    public SakuraException(String message) {
-        super(message);
-    }
-}
